backend/requirements.txt
fastapi[all]==0.95.2
uvicorn[standard]==0.21.1
alpaca-trade-api==3.0.0
SQLAlchemy==2.0.19
asyncpg==0.27.0
aioredis==2.0.1
pydantic==1.10.11
httpx==0.24.1
pandas==2.1.2
numpy==1.26.2
vectorbt==0.24.1
python-dotenv==1.0.0
apscheduler==3.10.1
